Ensure trailing slash for customize-preview-js src
On a multisite subdirectory install, for a site 'foo', a URL
like the following results in a NOBLOGREDIRECT:

http://example.com/foo?styles-action=customize-preview-js...

This results in the preview not working at all. The fix is to ensure
that the script path ends in a slash, and this helps WordPress Multisite
properly route the request to the current site.