Refactor control add_item, allow section overrides
By refactoring the logic to construct the args passed into $wp_customize->add_setting
and $wp_customize->add_control, less duplicated logic is in the controls.
With the control args array now constructed in one place, this gives the
opportunity to introduce a new element field 'section' that can override
the default group that a control is assigned to. This is ideal if you want
the styles controls to be located within existing sections (which I do).